requirements.txt
django==4.2.4
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-filter
python-dotenv
pymysql==1.1.0
cryptography==41.0.3
bacdive==1.0.0
pillow